enable tests on windows, move posix tests to own file
