+++ pkg/osquery/runtime/runtime_posix_test.go
@@ -1,0 +1,210 @@
+//go:build !windows
+// +build !windows
+
+package runtime
+
+import (
+	"fmt"
+	"log/slog"
+	"os/exec"
+	"path/filepath"
+	"syscall"
+	"testing"
+	"time"
+
+	typesMocks "github.com/kolide/launcher/ee/agent/types/mocks"
+	"github.com/kolide/launcher/pkg/log/multislogger"
+	"github.com/kolide/launcher/pkg/threadsafebuffer"
+	"github.com/osquery/osquery-go"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/mock"
+	"github.com/stretchr/testify/require"
+)
+
+func TestRestart(t *testing.T) {
+	t.Parallel()
+	runner, teardown := setupOsqueryInstanceForTests(t)
+	defer teardown()
+
+	previousStats := runner.instance.stats
+
+	require.NoError(t, runner.Restart())
+	waitHealthy(t, runner)
+
+	require.NotEmpty(t, runner.instance.stats.StartTime, "start time should be set on latest instance stats after restart")
+	require.NotEmpty(t, runner.instance.stats.ConnectTime, "connect time should be set on latest instance stats after restart")
+
+	require.NotEmpty(t, previousStats.ExitTime, "exit time should be set on last instance stats when restarted")
+	require.NotEmpty(t, previousStats.Error, "stats instance should have an error on restart")
+
+	previousStats = runner.instance.stats
+
+	require.NoError(t, runner.Restart())
+	waitHealthy(t, runner)
+
+	require.NotEmpty(t, runner.instance.stats.StartTime, "start time should be added to latest instance stats after restart")
+	require.NotEmpty(t, runner.instance.stats.ConnectTime, "connect time should be added to latest instance stats after restart")
+
+	require.NotEmpty(t, previousStats.ExitTime, "exit time should be set on instance stats when restarted")
+	require.NotEmpty(t, previousStats.Error, "stats instance should have an error on restart")
+}
+
+// TestExtensionIsCleanedUp tests that the osquery extension cleans
+// itself up. Unfortunately, this test has proved very flakey on
+// circle-ci, but just fine on laptops.
+func TestExtensionIsCleanedUp(t *testing.T) {
+	t.Skip("https://github.com/kolide/launcher/issues/478")
+	t.Parallel()
+
+	runner, teardown := setupOsqueryInstanceForTests(t)
+	defer teardown()
+
+	osqueryPID := runner.instance.cmd.Process.Pid
+
+	pgid, err := syscall.Getpgid(osqueryPID)
+	require.NoError(t, err)
+	require.Equal(t, pgid, osqueryPID, "pgid must be set")
+
+	require.NoError(t, err)
+
+	// kill the current osquery process but not the extension
+	err = syscall.Kill(osqueryPID, syscall.SIGKILL)
+	require.NoError(t, err)
+
+	// We need to (a) let the runner restart osquery, and (b) wait for
+	// the extension to die. Both of these may take up to 30s. We'll
+	// start a clock, wait for the respawn, and after 32s, test that the
+	// extension process is no longer running. See
+	// https://github.com/kolide/launcher/pull/342 and associated for
+	// background.
+	timer1 := time.NewTimer(35 * time.Second)
+
+	// Wait for osquery to respawn
+	waitHealthy(t, runner)
+
+	// Ensure we've waited at least 32s
+	<-timer1.C
+}
+
+func TestOsquerySlowStart(t *testing.T) {
+	t.Parallel()
+	rootDirectory, rmRootDirectory, err := osqueryTempDir()
+	require.NoError(t, err)
+	defer rmRootDirectory()
+
+	var logBytes threadsafebuffer.ThreadSafeBuffer
+
+	k := typesMocks.NewKnapsack(t)
+	k.On("OsqueryHealthcheckStartupDelay").Return(0 * time.Second).Maybe()
+	k.On("WatchdogEnabled").Return(false)
+	k.On("RegisterChangeObserver", mock.Anything, mock.Anything, mock.Anything, mock.Anything, mock.Anything)
+	slogger := multislogger.New(slog.NewJSONHandler(&logBytes, &slog.HandlerOptions{Level: slog.LevelDebug}))
+	k.On("Slogger").Return(slogger.Logger)
+	k.On("PinnedOsquerydVersion").Return("")
+
+	runner := New(
+		k,
+		WithKnapsack(k),
+		WithRootDirectory(rootDirectory),
+		WithOsquerydBinary(testOsqueryBinaryDirectory),
+		WithSlogger(slogger.Logger),
+		WithStartFunc(func(cmd *exec.Cmd) error {
+			err := cmd.Start()
+			if err != nil {
+				return fmt.Errorf("unexpected error starting command: %w", err)
+			}
+			// suspend the process right away
+			cmd.Process.Signal(syscall.SIGTSTP)
+			go func() {
+				// wait a while before resuming the process
+				time.Sleep(3 * time.Second)
+				cmd.Process.Signal(syscall.SIGCONT)
+			}()
+			return nil
+		}),
+	)
+	go runner.Run()
+	waitHealthy(t, runner)
+
+	// ensure that we actually had to wait on the socket
+	require.Contains(t, logBytes.String(), "osquery extension socket not created yet")
+	require.NoError(t, runner.Shutdown())
+}
+
+func TestExtensionSocketPath(t *testing.T) {
+	t.Parallel()
+
+	rootDirectory, rmRootDirectory, err := osqueryTempDir()
+	require.NoError(t, err)
+	defer rmRootDirectory()
+
+	k := typesMocks.NewKnapsack(t)
+	k.On("OsqueryHealthcheckStartupDelay").Return(0 * time.Second).Maybe()
+	k.On("WatchdogEnabled").Return(false)
+	k.On("RegisterChangeObserver", mock.Anything, mock.Anything, mock.Anything, mock.Anything, mock.Anything)
+	k.On("Slogger").Return(multislogger.NewNopLogger())
+	k.On("PinnedOsquerydVersion").Return("")
+
+	extensionSocketPath := filepath.Join(rootDirectory, "sock")
+	runner := New(
+		k,
+		WithKnapsack(k),
+		WithRootDirectory(rootDirectory),
+		WithExtensionSocketPath(extensionSocketPath),
+		WithOsquerydBinary(testOsqueryBinaryDirectory),
+	)
+	go runner.Run()
+
+	waitHealthy(t, runner)
+
+	// wait for the launcher-provided extension to register
+	time.Sleep(2 * time.Second)
+
+	client, err := osquery.NewClient(extensionSocketPath, 5*time.Second, osquery.DefaultWaitTime(1*time.Second), osquery.MaxWaitTime(1*time.Minute))
+	require.NoError(t, err)
+	defer client.Close()
+
+	resp, err := client.Query("select * from launcher_gc_info")
+	require.NoError(t, err)
+	assert.Equal(t, int32(0), resp.Status.Code)
+	assert.Equal(t, "OK", resp.Status.Message)
+
+	require.NoError(t, runner.Shutdown())
+}
+
+// sets up an osquery instance with a running extension to be used in tests.
+func setupOsqueryInstanceForTests(t *testing.T) (runner *Runner, teardown func()) {
+	rootDirectory, rmRootDirectory, err := osqueryTempDir()
+	require.NoError(t, err)
+
+	k := typesMocks.NewKnapsack(t)
+	k.On("OsqueryHealthcheckStartupDelay").Return(0 * time.Second).Maybe()
+	k.On("WatchdogEnabled").Return(true)
+	k.On("WatchdogMemoryLimitMB").Return(150)
+	k.On("WatchdogUtilizationLimitPercent").Return(20)
+	k.On("WatchdogDelaySec").Return(120)
+	k.On("RegisterChangeObserver", mock.Anything, mock.Anything, mock.Anything, mock.Anything, mock.Anything).Maybe()
+	k.On("Slogger").Return(multislogger.NewNopLogger())
+	k.On("PinnedOsquerydVersion").Return("")
+
+	runner = New(
+		k,
+		WithKnapsack(k),
+		WithRootDirectory(rootDirectory),
+		WithOsquerydBinary(testOsqueryBinaryDirectory),
+	)
+	go runner.Run()
+	waitHealthy(t, runner)
+
+	osqueryPID := runner.instance.cmd.Process.Pid
+
+	pgid, err := syscall.Getpgid(osqueryPID)
+	require.NoError(t, err)
+	require.Equal(t, pgid, osqueryPID)
+
+	teardown = func() {
+		defer rmRootDirectory()
+		require.NoError(t, runner.Shutdown())
+	}
+	return runner, teardown
+}
